requirements.txt
tabulate==0.8.10
africastalking~=1.2.4
python-dotenv~=0.21.0